+++ tests/Main.hs
@@ -15,6 +15,7 @@ import qualified GetDirContents001
 import qualified GetDirContents002
 import qualified GetHomeDirectory001
 import qualified GetPermissions001
+import qualified IsSymbolicLink
 import qualified RemoveDirectoryRecursive001
 import qualified RenameFile001
 import qualified Safe
@@ -38,6 +39,7 @@ main = T.testMain $ \ _t -> do
   T.isolatedRun _t "GetDirContents002" GetDirContents002.main
   T.isolatedRun _t "GetHomeDirectory001" GetHomeDirectory001.main
   T.isolatedRun _t "GetPermissions001" GetPermissions001.main
+  T.isolatedRun _t "IsSymbolicLink" IsSymbolicLink.main
   T.isolatedRun _t "RemoveDirectoryRecursive001" RemoveDirectoryRecursive001.main
   T.isolatedRun _t "RenameFile001" RenameFile001.main
   T.isolatedRun _t "Safe" Safe.main