+++ cbits/directory.c
@@ -1,5 +1,4 @@
-#if defined(__GLASGOW_HASKELL__) || defined(__HUGS__)
-/* 
+/*
  * (c) The University of Glasgow 2002
  *
  */
@@ -8,6 +7,3 @@
 
 #define INLINE
 #include "HsDirectory.h"
-
-#endif
-