+++ changelog.md
@@ -15,6 +15,8 @@ Changelog for the [`directory`][1] package
   * Improve error message of `removeDirectoryRecursive` when used on a
     directory symbolic link on Windows.
 
+  * Add `isSymbolicLink`
+
 ## 1.2.5.1 (February 2015)
 
   * Improve error message of `getCurrentDirectory` when the current working