+++ changelog.md
@@ -3,6 +3,8 @@ Changelog for the [`directory`][1] package
 
 ## 1.2.6.0 (April 2015)
 
+  * Bundled with GHC 8.0.1
+
   * Make `findExecutable`, `findExecutables`, `findExecutablesInDirectories`,
     `findFile`, and `findFilesWith` lazier
     ([#43](https://github.com/haskell/directory/issues/43))